Bug 22988461 - INCORRECT CALCULATION FOR PAGES_HASHED, PAGES_OLD
The views:

   * innodb_buffer_stats_by_schema
   * innodb_buffer_stats_by_table

used COUNT(IF(...), 1, 0) to sum up the number of hashed and old pages.
However COUNT(0) will increment the counter, so the counts have been changed
to COUNT(IF(...), 1, NULL) as NULL will not increment the counter. This solution
is preferred over SUM(IF(...), 1, 0) as SUM returns a decimal instead of an
integer.

Also added test cases to ensure the percentage of old pages found is less
than @@global.innodb_old_blocks_pct as a check that not all pages are included
in the count.