Bug #21512106 FORMAT_PATH() DOES NOT CONSIDER DIRECTORY BOUNDARIES
Properly check directory boundaries on all variables now. If the variable does not include a path separator at the end of it, add one for comparison.

NOTE: This stops the case of converting Windows paths to /, and now correctly shows \ for Windows hosts.

Also added replacement for the basedir variable, and added a proper unit test that validates all variables are parsed correctly (though this requires disabling the test on Windows, as the path separator is now correctly shown as \ on that host too now).