Add is_system_schema() function
Implements this RFE:
Bug #78314 	Register if a schema is a system schema or not
https://bugs.mysql.com/bug.php?id=78314

Note that a hardcoded list of system schemas will make it difficult
to later add/remove schemas. It would be better if a schema could register
itself as a system schema (ALTER SCHEMA foo IS SYSTEM SCHEMA) or by
modifying a table like sys.system_schema_list

Would be more useful if we could do SHOW USER SCHEMAS / SHOW SYSTEM SCHEMAS

Not including the results file for the test case.